requirements.txt
jupyter_kernel_gateway
mysql-connector-python
jupyter-server-proxy
flask-mysql
